refactor: 座標バリデーション・丸め処理を coordinateValidator に統一
