+++ src/lib/validators/index.ts
@@ -1,1 +1,2 @@
 export { isValidUUID, UUID_REGEX } from './uuidValidator';
+export { roundCoordinate, parseAndValidateCoordinates } from './coordinateValidator';
